refactored plugins
Functionality is the same but architecture has changed a bit to prepare merging definition based metric groups and samples into the Metric namespace and simplify the core's sampling code.
Another target is to prepare the code for running as a daemon without having to parse source code on every sampling.