looped sampling allows memorization of sample values
Instead of forking a thread is started. This allows the sampling code to memorize values with
metric's attr_accessors.
After each run the garbage collection is started to remove the dead thread and all unreferenced
memory.
Potentially this is not as clean as the forking solution but hopefully it will be okay.